Migrate prompts to dedicated folder
